equipo3Pedidos/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
mysqlclient
python-decouple
requests
